// frontend/react-template/src/models/CustomerInfoModel.ts
export interface CustomerInfoModel {
  id: string;
  name: string;
  gender: string;
  idNumber: string;
  dayOfBirth: string;
  familyRelation: string;
  agency: string;
  contractNumber: string;
  companyName: string;
  companyAddress: string;
  companyDistrict: string;
  companyProvince: string;
  group: string;
}